+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# [WIP 0.17.2](https://github.com/upb-uc4/ui-web/compare/v0.17.1-hotfix.2...v0.17.2) (2021-02-0X)
+## Feature
+- add new operations workflow [#832](https://github.com/upb-uc4/ui-web/pull/832)
+- add ability to watch operations
+
+## Bugfix
+- fix colour of uc4 logo in mobile navigation [#834](https://github.com/upb-uc4/ui-web/pull/834)
+
+# [0.17.1-hotfix.2](https://github.com/upb-uc4/ui-web/compare/v0.17.1-hotfix.1...v0.17.1-hotfix.2) (2021-02-03)
+## Bugfix
+- Fix a bug that broke matriculation
+
 # [0.17.1-hotfix.1](https://github.com/upb-uc4/ui-web/compare/v0.17.1...v0.17.1-hotfix.1) (2021-02-03)
 ## Bugfix
 - Fix purgecss removing our fancy dark mode background. #828 [#829](https://github.com/upb-uc4/ui-web/pull/829)